Fix parsing issues for PCRE error messages
The PCRE error messages have slightly different format (no whitespace
between the different fields inclosed in square brackets)

This leads to parsing issues in the awk script, as it doesn't recognize
the fields correctly anymore due to the missing whitespace.

By adjusting the input so that between a closing and opening square
braken there always is a whitespace, we ensure that the fields are
correctly distinguished.

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>