Set initial value for in- and outbound anomaly score
The anomaly score variables are not set to a specific value at the
beginning. This can lead to an empty field in the access-log if a rule
is triggered with a `deny`. The reason is, that a deny will abort any
further rule evaluation and therefore also the anomaly score
calculation. Initializing the scores to 0 at the beginning eliminates
this issue.

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>